refactor: move `fraud` to `libs/fraud`
