+++ libs/fraud/helpers.go
