+++ libs/fraud/interface.go
