+++ libs/fraud/proof.go
