+++ libs/fraud/registry.go
