+++ nodebuilder/fraud/mocks/api.go
@@ -10,8 +10,8 @@ import (
 
 	gomock "github.com/golang/mock/gomock"
 
-	fraud "github.com/celestiaorg/celestia-node/fraud"
 	fraud0 "github.com/celestiaorg/celestia-node/nodebuilder/fraud"
+	"github.com/celestiaorg/celestia-node/libs/fraud"
 )
 
 // MockModule is a mock of Module interface.
