+++ nodebuilder/p2p/pubsub.go
@@ -15,7 +15,7 @@ import (
 	"go.uber.org/fx"
 	"golang.org/x/crypto/blake2b"
 
-	"github.com/celestiaorg/celestia-node/fraud"
+	"github.com/celestiaorg/celestia-node/libs/fraud"
 	headp2p "github.com/celestiaorg/celestia-node/libs/header/p2p"
 )
